small adjustments - fixing some texts - adding a small tooltip - renaming some files and texts
