+++ app/components/markdown/markdown-editor.tsx
@@ -35,8 +35,6 @@ export const MarkdownEditor = forwardRef(function MarkdownEditor(
   }: Props,
   ref
 ) {
-  // const sendMagicLink = useTypedFetcher<typeof action>();
-
   const fetcher = useFetcher<typeof action>();
   const content = fetcher?.data?.content;
   const [markdown, setMarkdown] = useAtom(markdownAtom);
