+++ app/routes/_auth+/login.tsx
@@ -217,7 +217,7 @@ export default function IndexLoginForm() {
           </div>
           <div className="relative flex justify-center text-sm">
             <span className="bg-white px-2 text-gray-500">
-              Or use a <strong>Magic Link</strong>
+              Or use a <strong>One Time Password</strong>
             </span>
           </div>
         </div>
