+++ server/index.ts
@@ -102,9 +102,9 @@ app.use(
       "/logout",
       "/otp",
       "/oauth/callback",
-      "/resend-email-confirmation",
+      "/resend-otp",
       "/reset-password",
-      "/send-magic-link",
+      "/send-otp",
       "/verify-email",
       "/healthcheck",
       "/api/public-stats",
